Limit ObjectMessage deserialization with a package white list
ObjectMessages in JMS can be used to transfer arbitrary Java objects
which are then deserialized before or at the time of
being passed on to message listeners.

This opens message consumers up to a range of attacks
that exploit issues in Java object serialization.
Note: attacker must authenticate with RabbitMQ in order
to carry out the attack.

With this commit we introduce a way to white list packages
that can be deserialized when ObjectMessages are consumed:

 * Using a system property, com.rabbitmq.jms.TrustedPackagesPrefixes
 * Using RMQSession#setTrustedPackages(List<String>)

By default all packages are trusted for backwards compatibility
(as it's not possible for library maintainers to cover application-specific
packages).

A CVE has been allocated for this issue and will be announced
once a new release is published.

Related issues in other messaging libraries and services:

 * Spring AMQP ticket: https://jira.spring.io/browse/AMQP-590
 * ActiveMQ ticket: https://issues.apache.org/jira/browse/AMQ-6013

Note: this includes a number of small drive-by changes:

 * TLS tests are now executed conditionally and excluded by default
 * Many minor IDEA code analysis fixes
 * JavaDoc warning fixes