manage_test_broker.groovy: Use `rabbitmq_jms_topic_exchange` directly
... instead of `rabbit` and its `PLUGINS` variable.

The latter abuses a facility leading to circular dependencies. Now that
those cycles are being fixed, this causes an infinite loop: rabbit and
rabbitmq_jms_topic_exchange depending on each other.

There is no need to use the `PLUGINS` variable here as we only want to
start the broker with a single plugin enabled. So use the correct way to
do this: run `make start-background-broker` from the plugin's directory
directly.

[#130659985]