Take message properties into account on sending
Priority, delivery mode, and expiration aren't taken into
account on sending when set in the JMS message. This is compliant with
the specifiation though. This commit provide a flag in
RMQConnectionFactory to make the RMQMessageProducer use the
message properties if set. This is non-compliant with the specification
but more natural. The existing and spec-compliant behavior is still
the default.

Fixes #26