Add option to requeue message on client runtime exception
Requeuing a message after a client RuntimeException complies to
the JMS specification. Nevertheless, the default behavior is
still the same: the message is "lost" (as it has been AMQP-acknowledged
before the delivery to the client) if the client fails to process
it (which makes more sense to us).

The option is RMQConnectionFactory#requeueOnMessageListenerException
and is false by default.

Fixes #23