set the auto-delete to true for broker queues created to support non-durable topics with autogenerated names
Broker queues defined for Topic consumers of non-durable Topics are assigned auto-generated names that are not visible to JMS clients.

These broker queues are inherently transient and were previously (correctly) unbounded when the connection to the broker was closed.  However, because these broker queues are created per consumer, the broker would effectively leak these until the connection was closed.

This change flags such broker queues with the auto-delete flag, enabling the broker to expunge them once the last consumer is closed.