Handle request-reply with RabbitMQ direct reply-to
This commit provides support for RabbitMQ direct reply-to when
performing RPC operations. The client just needs to set up the
JMSReplyTo header with a RMQDestination named "amq.rabbitmq.reply-to".

This provides better performance than using a temporary queue, but it
comes with 2 limitations:
 * the reply-to destination is an JMS "AMQP" destination, so only
ByteMessages and TextMessages are supported.
 * the client needs to create a MessageConsumer and register a
MessageListener to get the response, as direct reply-to doesn't work
with basic.get.

Fixes #48