Do not declare reply-to destination for consumed messages
This commit instructs to not declare a reply-to RMQDestination for
consumed messages. This avoids trying to create a second time a
temporary reply-to destination (reply-to destinations are supposed to be
created by a RPC client before sending the request).

This happens to also fix #47.

Fixes #69

(cherry picked from commit 09282e4a6e33c044ed045fd0a189f306d245c35e)

Conflicts:
	src/test/java/com/rabbitmq/integration/tests/RpcIT.java
	src/test/java/com/rabbitmq/integration/tests/RpcWithAmqpDirectReplyIT.java