Add callback before receiving message
This commit adds a callback on message consumption, after the AMQP
message is converted into a JMS message, but before it is dispatched to
application code.

This is used internally to ask to not declare the reply-to queue of a
received message. A previous commit would make this change every time,
but this broker the compliance test suite. So this is now an opt-in,
useful for the server-side of RPC use cases, when we don't want to
re-create the temporary reply-to destination.

References #69