Return 0 JMS timestamp if property is null
The property is not supposed to be null in JMS, but it can in AMQP
0.9.1, so JMS messages consumed from AMQP resources can have a null
value. This commit mitigates the issue by returning a timestamp set to 0
in this case, just like the JMS 1.1 specification suggests it when
MessageProducer#setDisableMessageTimestamp(true) is called.

Fixes #126

(cherry picked from commit 0fcf4c1c966f8c74715bc9469cf79c60d5a13d81)

Conflicts:
	src/test/java/com/rabbitmq/integration/tests/SimpleAmqpQueueMessageIT.java