Add autoJmsTypeHeaderForTextMessages property
When this option is enabled, the JMSType AMQP header is
set to TextMessage automatically for JMS TextMessages published
to AMQP resources. This way those messages are directly converted
to TextMessages for JMS consumers (instead of BytesMessages).

This avoids changing application code for some use cases: consumers
do not have to do some painful conversion to read the text body from
BytesMessages, they get TextMessages directly. And publishers
do not have to worry about setting the header depending on the
type of destination ("pure" JMS or AMQP).

The property is also settable from JNDI, which allows changing the
types of destinations without changing application code.

The option is disabled by default, so backward compatibility is preserved.