Clean pom.xml
Remove the profile to start up a node from the umbrella, because
it's not used anywhere now.

Simplify TLS testing, loading the CA certificate in the hostname
verification test, which avoids using intermediate keystore. So
remove the generation of these keystores (no more keytool plugin
and Groovy scripts).

Use only one Javadoc plugin configuration (not repeated in the
deployment plugins).