Support JMS 2.0 asynchronous send
Enable publisher confirms as soon as one of the MessageProducer
send methods with a CompletionListener parameter is used.
This will enable publisher confirms on all the MessageProducer instances
the owning session created or will create, because they share
the same AMQP channel.

Messages sent with the "sync" methods will use automatically
an empty CompletionListener.

The ConfirmListener API is now deprecated in favor of the async send
methods and the 2 mechanisms should not be used at the same time.

References #180

Fixes #205, #206