Support client ID set/get in JMSContext
The client ID must be set right after the connection creation,
so before the connection can even create a session.
This makes the state of a JMSContext hard to create, because
the session must then be created lazily (instead of being
able to use a final property for it).

References #180, #210