Add pluggable strategy for dealing with reply to queues.
Add a pluggable strategy for handling reply to queues when
messages are received with the a reply to specified.

This will default to the standard behaviour of just handling
"amq.rabbitmq.reply-to" destinations.

An alternative stategy, ReturnToSenderExchange, has also been
implemented that assumes that reply to queues, if not
"amq.rabbitmq.reply-to" are returned on the same exchange as the
received message.

The reply to strategy can be set on the connection factory.

(cherry picked from commit b027201c91c30818ce292da3a3819f1e48838b2e)

Conflicts:
	src/main/java/com/rabbitmq/jms/admin/RMQConnectionFactory.java
	src/main/java/com/rabbitmq/jms/client/ConnectionParams.java